# docs/api_docs.yaml
components:
  schemas:
...
    Cors:
      properties:
        allow_credentials:
          title: Allow Credentials
          type: boolean
        allow_headers:
          items:
            type: string
          title: Allow Headers
          type: array
        allow_methods:
          items:
            type: string
          title: Allow Methods
          type: array
        allow_origins:
          items:
            type: string
          title: Allow Origins
          type: array
        expose_headers:
          items:
            type: string
          title: Expose Headers
          type: array
      required:
      - allow_origins
      - allow_credentials
      - allow_methods
      - allow_headers
      - expose_headers
      title: Cors
      type: object
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    HttpServer:
      properties:
        description:
          title: Description
          type: string
        name:
          title: Name
          type: string
        tools:
          items:
            type: string
          title: Tools
          type: array
        url:
          title: Url
          type: string
      required:
      - name
      - description
      - url
      - tools
      title: HttpServer
      type: object
...
    Mcp:
      properties:
        created_at:
          format: date-time
          title: Created At
          type: string
        deleted_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Deleted At
        http_servers:
          items:
            $ref: '#/components/schemas/HttpServer'
          title: Http Servers
          type: array
        name:
          title: Name
          type: string
        routers:
          items:
            $ref: '#/components/schemas/Router'
          title: Routers
          type: array
        servers:
          items:
            $ref: '#/components/schemas/McpServer'
          title: Servers
          type: array
        tenant_name:
          title: Tenant Name
          type: string
        tools:
          items:
            $ref: '#/components/schemas/Tool'
          title: Tools
          type: array
        updated_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Updated At
      required:
      - name
      - tenant_name
      - updated_at
      - created_at
      - deleted_at
      - servers
      - routers
      - tools
      - http_servers
      title: Mcp
      type: object
    McpConfigModel:
      properties:
        gmt_created:
          format: date-time
          title: Gmt Created
          type: string
        gmt_deleted:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Gmt Deleted
        gmt_updated:
          format: date-time
          title: Gmt Updated
          type: string
        http_servers:
          items:
            $ref: '#/components/schemas/HttpServer'
          title: Http Servers
          type: array
        id:
          title: Id
          type: string
        name:
          title: Name
          type: string
        routers:
          items:
            $ref: '#/components/schemas/Router'
          title: Routers
          type: array
        servers:
          items:
            $ref: '#/components/schemas/McpServer'
          title: Servers
          type: array
        tenant_name:
          title: Tenant Name
          type: string
        tools:
          items:
            $ref: '#/components/schemas/Tool'
          title: Tools
          type: array
      required:
      - id
      - name
      - tenant_name
      - routers
      - servers
      - tools
      - http_servers
      - gmt_created
      - gmt_updated
      title: McpConfigModel
      type: object
    McpConfigName:
      description: MCP config name.
      properties:
        id:
          title: Id
          type: string
        name:
          title: Name
          type: string
        tenant_name:
          title: Tenant Name
          type: string
      required:
      - id
      - name
      - tenant_name
      title: McpConfigName
      type: object
    McpServer:
      properties:
        args:
          items:
            type: string
          title: Args
          type: array
        command:
          title: Command
          type: string
        description:
          title: Description
          type: string
        name:
          title: Name
          type: string
        policy:
          $ref: '#/components/schemas/Policy'
        preinstalled:
          title: Preinstalled
          type: boolean
        type:
          $ref: '#/components/schemas/McpServerType'
        url:
          title: Url
          type: string
      required:
      - name
      - type
      - description
      - policy
      - command
      - preinstalled
      - url
      - args
      title: McpServer
      type: object
    McpServerType:
      enum:
      - sse
      - stdio
      title: McpServerType
      type: string
...
    Policy:
      enum:
      - on_start
      - on_demand
      title: Policy
      type: string
...
    Router:
      properties:
        cors:
          anyOf:
          - $ref: '#/components/schemas/Cors'
          - type: 'null'
        prefix:
          title: Prefix
          type: string
        server:
          title: Server
          type: string
        sse_prefix:
          title: Sse Prefix
          type: string
      required:
      - prefix
      - server
      - sse_prefix
      title: Router
      type: object
...
    Tool:
      properties:
        args:
          items:
            additionalProperties: true
            type: object
          title: Args
          type: array
        description:
          title: Description
          type: string
        headers:
          additionalProperties:
            type: string
          title: Headers
          type: object
        input_schema:
          additionalProperties: true
          title: Input Schema
          type: object
        method:
          title: Method
          type: string
        name:
          title: Name
          type: string
        path:
          title: Path
          type: string
        request_body:
          title: Request Body
          type: string
        response_body:
          title: Response Body
          type: string
      required:
      - name
      - description
      - method
      - path
      - headers
      - args
      - request_body
      - response_body
      - input_schema
      title: Tool
      type: object
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
  securitySchemes:
    APIKeyCookie:
      in: cookie
      name: session
      type: apiKey
info:
  title: API Server
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /api/v1/mcp/configs:
    get:
      description: 获取MCP配置列表
      operationId: list_mcp_configs_api_v1_mcp_configs_get
      parameters:
      - description: 租户ID
        in: query
        name: tenant_name
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 租户ID
          title: Tenant Name
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/McpConfigModel'
                title: Response List Mcp Configs Api V1 Mcp Configs Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - APIKeyCookie: []
      summary: List Mcp Configs
    post:
      description: 创建MCP配置
      operationId: create_mcp_config_api_v1_mcp_configs_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Mcp'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/McpConfigModel'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - APIKeyCookie: []
      summary: Create Mcp Config
    put:
      description: 更新MCP配置
      operationId: update_mcp_config_api_v1_mcp_configs_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Mcp'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/McpConfigModel'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - APIKeyCookie: []
      summary: Update Mcp Config
  /api/v1/mcp/configs/names:
    get:
      description: 获取MCP配置名称列表
      operationId: list_mcp_config_names_api_v1_mcp_configs_names_get
      parameters:
      - description: 租户ID
        in: query
        name: tenant_name
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 租户ID
          title: Tenant Name
      - description: 是否包含已删除的配置
        in: query
        name: include_deleted
        required: false
        schema:
          default: false
          description: 是否包含已删除的配置
          title: Include Deleted
          type: boolean
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/McpConfigName'
                title: Response List Mcp Config Names Api V1 Mcp Configs Names Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - APIKeyCookie: []
      summary: List Mcp Config Names
  /api/v1/mcp/configs/{config_id}/sync:
    post:
      description: 同步MCP配置
      operationId: sync_mcp_config_api_v1_mcp_configs__config_id__sync_post
      parameters:
      - in: path
        name: config_id
        required: true
        schema:
          title: Config Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - APIKeyCookie: []
      summary: Sync Mcp Config
  /api/v1/mcp/configs/{tenant_name}/{name}:
    delete:
      description: 删除MCP配置
      operationId: delete_mcp_config_api_v1_mcp_configs__tenant_name___name__delete
      parameters:
      - in: path
        name: tenant_name
        required: true
        schema:
          title: Tenant Name
          type: string
      - in: path
        name: name
        required: true
        schema:
          title: Name
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - APIKeyCookie: []
      summary: Delete Mcp Config
  /api/v1/mcp/{tenant_name}/{name}/active:
    post:
      description: 激活MCP配置
      operationId: active_mcp_config_api_v1_mcp__tenant_name___name__active_post
      parameters:
      - in: path
        name: tenant_name
        required: true
        schema:
          title: Tenant Name
          type: string
      - in: path
        name: name
        required: true
        schema:
          title: Name
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - APIKeyCookie: []
      summary: Active Mcp Config
...